chore: extract the "book a call" section into components/pages/pricing
